pyproject.toml
[project]
name = "dotfiles"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gsy0911", email = "yoshiki0911@gmail.com" }
]
dependencies = [
    "pydantic>=2.10.5",
    "pyyaml>=6.0.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
default-groups = "all"
managed = true

[dependency-groups]
dev = [
    "marimo>=0.19.2",
    "ruff>=0.9.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dotfiles"]
